requirements.txt
# Knowledge-Hub — Python dependencies
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# DB layer
sqlalchemy==2.0.48
asyncpg==0.29.0
pgvector==0.3.6

# Auth — OIDC/JWT (D03: PyJWT >= 2.8)
PyJWT>=2.8.0
cryptography>=42.0.0
httpx>=0.27.0

# CJK tokenizers (cjk-tokenizer feature)
mecab-python3>=1.0.6
kiwipiepy>=0.18.0
jieba>=0.42.1
underthesea>=6.8.0
langdetect>=1.0.9

# LLM provider adapters (llm-provider feature)
openai>=1.0.0
anthropic>=0.20.0

# Rate limiting — Valkey client (AC3: use valkey-py, not redis-py)
valkey>=6.0.0

# Testing
pytest==8.3.5
pytest-asyncio==0.25.3
